chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-time-range-expected.txt

This is a testharness.js-based test.
FAIL Timerange mapped correctly (10s) assert_equals: expected "150px" but got "100px"
FAIL Timerange mapped correctly (1s) assert_equals: expected "105px" but got "100px"
FAIL Timerange mapped correctly (1000ms) assert_equals: expected "105px" but got "100px"
Harness: the test ran to completion.
